#16 Cast numbers to integers for math ops

+++ widgets/edd-widget-downloads-calendar.php
@@ -292,8 +292,8 @@ if ( ! class_exists( 'EDD_Downloads_Calendar' ) ) {
 				}
 			}
 
-			// See how much we should pad in the beginning.
-			$pad                  = calendar_week_mod( date( 'w', $unixmonth ) - $week_begins );
+			// See how much we should pad at the beginning of the month.
+			$pad = (int) calendar_week_mod( date( 'w', $unixmonth ) - $week_begins );
 			if ( 0 !== $pad ) {
 				$calendar_output .= "\n\t\t" . '<td colspan="' . esc_attr( $pad ) . '" class="pad">&nbsp;</td>';
 			}
@@ -317,12 +317,13 @@ if ( ! class_exists( 'EDD_Downloads_Calendar' ) ) {
 				}
 				$calendar_output .= '</td>';
 
-				if ( 6 === calendar_week_mod( date( 'w', mktime( 0, 0, 0, $thismonth, $day, $thisyear ) ) - $week_begins ) ) {
+				if ( 6 === (int) calendar_week_mod( date( 'w', mktime( 0, 0, 0, $thismonth, $day, $thisyear ) ) - $week_begins ) ) {
 					$newrow = true;
 				}
 			}
 
-			$pad                  = 7 - calendar_week_mod( date( 'w', mktime( 0, 0, 0, $thismonth, $day, $thisyear ) ) - $week_begins );
+			// See how much we should pad at the end of the month.
+			$pad = 7 - (int) calendar_week_mod( date( 'w', mktime( 0, 0, 0, $thismonth, $day, $thisyear ) ) - $week_begins );
 			if ( 0 !== $pad && 7 !== $pad ) {
 				$calendar_output .= "\n\t\t" . '<td class="pad" colspan="' . esc_attr( $pad ) . '">&nbsp;</td>';
 			}
